Refactor VhostsClient and add Get method for RabbitMQ management
